Redo schema and data

+++ core/schema.sql
@@ -24,7 +24,7 @@ CREATE TABLE IF NOT EXISTS purchase (
     customer_id INTEGER NOT NULL, -- foreign key referencing the customer who made the purchase
     product_id INTEGER NOT NULL, -- foreign key referencing the product that was purchased
     quantity INTEGER NOT NULL CHECK(quantity > 0), -- quantity of the product purchased, must be positive
-    total_price INTEGER NOT NULL CHECK(total_price >= 0), -- total price of the purchase, must be non-negative
+    total_price REAL NOT NULL CHECK(total_price >= 0), -- total price of the purchase, must be non-negative
     purchase_date DATE NOT NULL -- date of the purchase, cannot be empty,...
             DEFAULT CURRENT_DATE -- defaults to current date if not provided, and...
             CHECK(purchase_date <= CURRENT_DATE), -- cannot be in the future
